==> Code/01-data-joining.Rmd <==
---
title: "Join zip data to values data"
date: "19 11 2018"
output: html_document
params:
  date: !r Sys.Date()
  verbose: TRUE
editor_options: 
  chunk_output_type: console
---


# Setup


```{r knitr-setup}
knitr::opts_chunk$set(
  comment = "#>",
  collapse = TRUE,
  message = FALSE,
  warning = FALSE,
  cache = TRUE,
  out.width = "70%",
  fig.align = 'center',
  fig.width = 6,
  fig.asp =  0.4,  #0.618,  # 1 / phi
  fig.show = "hold",
  size = "tiny"
)

```



## Load packages


```{r}
library(sjmisc)
library(viridis)
library(pradadata)  # elec data
library(tidyverse)
library(knitr)
library(sf)

library(rethinking)  # Bayes modeling
library(rprojroot)
```




```{r}
proj_root <- rprojroot::find_root(as.root_criterion(is_rstudio_project))
```





# Load data



## Zip data
Load zip data (PLZ):


```{r load-zip-data}
zip <- readxl::read_excel(paste0(proj_root, "/data_raw/Wahlkreise_PLZ.xlsx"))
```


This dataset was accessed form [DeStatis](https://www.destatis.de/DE/ZahlenFakten/LaenderRegionen/Regionales/Gemeindeverzeichnis/Administrativ/Archiv/GVAuszugQ/BTW20174Q2016.html)


Copyright Notice:

 Daten (im Auftrag der Herausgebergemeinschaft Statistische Ämter des Bundes und der Länder)
      Statistisches Bundesamt, Wiesbaden 2017
           Vervielfältigung und Verbreitung, auch auszugsweise, mit Quellenangabe gestattet.
           
          
          
          
## Values data

### Load values data


```{r load-values-data}
load(paste0(proj_root, "/data_raw/Choughs2017.Rdata"))
values <- Choughs2017
```

The variable `fit.cluster` shows the seven different value types:

```{r}
values$fit.cluster %>% unique()
```




## Election data

These data are provided by the Bundeswahlleiter 2017.


```{r data-elec}
data("elec_results")
```


In order to combine socioeconomic data with the election results, we can make use of data from the same source as above. Again accessible via the same R pacakge:

```{r data-socec}
data("socec")
```

Note that a code book is available for these data:

```{r data-socec-dict}
data("socec_dict")
```



These data will be used as predictors for modeling the election results.

Third, we will make use of geo data in order to geoplot the modeling results. The Bundeswahlleiter provides such data (again via `pradadata`):

```{r data-shp}
data("wahlkreise_shp")
```



Note: Data objects can also be downloaded [from this source](https://osf.io/2yhr9/).



# Checks





## Check values data

Do the long and lat values appear plausible?


### Geo plot PLZ

```{r}
values %>% 
  select(lon, lat) %>% 
  gather( ) %>% 
  ggplot() +
  geom_jitter(aes(x = key, y = value), alpha = .7)
```


```{r}
values <- values %>% 
  filter(between(lon, 5, 15)) %>% 
  filter(between(lat, 45, 60)) %>% 
  drop_na(lon, lat)


values_sf <- values %>% 
  select(PLZ, lon, lat) %>% 
  st_as_sf(coords = c("lon", "lat"), crs = 4236)

values_sf

values_coord <- st_coordinates(values_sf) %>%
  cbind(values_sf)
```



```{r}
wahlkreise_shp %>% 
  ggplot() +
  geom_sf() +
  geom_point(data = values_coord,
             aes(x = X, y = Y)) +
    coord_sf(xlim = c(5, 15), ylim = c(45, 60))  


```



### Distrib of PLZ

```{r}
values %>% 
  mutate(PLZ_num = parse_number(PLZ)) %>% 
  ggplot(aes(x = PLZ_num)) +
  geom_histogram()
```


How many cases for PLZ are missing?

```{r}
values %>% 
  select(fit.cluster, PLZ ) %>% 
  drop_na(PLZ) %>% 
  nrow()
```


For how many ZIPs do we have value data?


```{r}
values %>% 
  distinct(PLZ) %>% 
  nrow()
  
```



## Check ZIP data


how many DIFFERENT PLZ values?

```{r}
unique(zip$PLZ) %>% length()
```


Ba-Wü?


```{r}
zip %>% 
  filter(str_detect(Gemeinde, "Stuttg"))

zip %>% 
  filter(str_detect(PLZ, "^70"))
```



```{r}
zip %>% 
  filter(str_detect(PLZ,"80336")) 


zip %>% 
  filter(str_detect(PLZ,"80636")) 

values %>% 
  filter(str_detect(PLZ,"80636")) 
```

In contrast, how many value data points (CHOUHGS) do we have alone for this PLZ?

```{r}
values %>% 
  filter(str_detect(PLZ,"8063")) %>% 
  nrow()

zip%>% 
  filter(str_detect(PLZ,"8063")) %>% 
  nrow()
```



Hamburg

```{r}
zip %>% 
  filter(str_detect(PLZ,"220")) %>% 
  nrow()
```


# How many ZIP can be joined to value?


```{r}
values %>% 
  left_join(zip) %>% 
  drop_na(PLZ) %>% 
  nrow()
```



And vice versA:

```{r}
zip %>% 
  left_join(values) %>% 
  drop_na(PLZ) %>% 
  nrow()
```


# Data joining


Now let's merge t

he data frames. There will also be some janitor work such as renaming columns etc.

## Prep

First, change the names of the `socec` data to a common format:

```{r soceec-rename}
socec_renamed <- socec %>%
  rename(state = V01,
         area_nr = V02,
         area_name = V03,
         total_n = V06,
         germans_n = V07,
         for_prop = V08,
         pop_move_prop = V11,
         pop_migr_background_prop = V19,
         income = V26,
         unemp_prop = V47) 

```


Compute some more columns and remove missing vlaues:

```{r socec-new-cols}
socec2 <- socec_renamed %>% 
   mutate(foreigner_n = total_n - germans_n,
         pop_move_n = pop_move_prop * total_n,
         unemp_n = unemp_prop * total_n / 100,
         pop_migr_background_n = pop_migr_background_prop * total_n / 100) %>% 
  drop_na()
```



Same thing with the election data, here we only need the criterion (AfD success) and the ID variables for merging:

```{r elec-results-new-cols}
elec_results2 <- elec_results %>%
  rename(afd_votes = AfD_3,
         area_nr = district_nr,
         area_name = district_name,
         votes_total = votes_valid_3) %>% 
   mutate(afd_prop = afd_votes/votes_total)    # valid votes only, and of the present Zweitstimme
```

Note that we are focusing on the Zweitstimme of the present election (hence the `3` in `votes_valid_3` and in `AfD_3`).

# Merge/Join socec data with election data


```{r merge}
socec2 %>%
  left_join(elec_results2, by = "area_name") %>% 
  left_join(wahlkreise_shp, by = c("area_name" = "WKR_NAME")) -> d_all_with_na
```


## After-merge preparations



Add variable for East (1) vs. West (0):


```{r east-west-dummy}
d_all_with_na <- d_all_with_na %>% 
  mutate(east = case_when(
    state %in% c("Mecklenburg-Vorpommern", "Brandenburg", "Berlin", "Sachsen-Anhalt", "Sachsen", "Thüringen") ~ "yes",
    TRUE ~ "no"
    ) 
  )

d_all_with_na$east_num <- ifelse(d_all_with_na$east == "yes", 1, 0)
```



## Main data frame: d_short 

We will also provide a version without the geo data, and in pure (old school) `data frame` form (ie., not as tibble)_

```{r d-short-with-nas}
d_all_with_na %>%
  rename(area_nr = area_nr.x) %>% 
  select(state,
         area_nr,
         area_name,
         total_n,
         germans_n,
         foreigner_n,
         for_prop,
         pop_move_n,
         pop_migr_background_n,
         income ,
         unemp_n,
         unemp_prop,
         votes_total,
         afd_votes,
         afd_prop,
         state,
         east,
         east_num) -> d_short_with_nas

if (params$verbose == TRUE) names(d_short_with_nas)
```


Remove NAs:

```{r nona}
d_short_with_nas %>% 
  drop_na() -> d_short_nona
```


Add state id:

```{r staste-id}
d_short_nona$state_id <- coerce_index(d_short_nona$state)

```


Create dictionnary for state id:

```{r state-dict} 
state_dict <- d_short_nona %>% 
  distinct(state_id, .keep_all = TRUE) %>% 
  select(state_id, state)
```





Multiply by 1000 to get the real numbers so that a count model gets the "right" data

```{r mult-by-1000}
d_short_nona %>%
  mutate_at(vars(total_n, germans_n, foreigner_n, pop_move_n,
                    pop_migr_background_n, unemp_n), funs(. * 1000)
  ) -> d_short_nona_1000

if (params$verbose == TRUE) glimpse(d_short_nona_1000)

```




# Join election data with zip data and values data LONG format



## Join main data with zip codes data


PLZ in zip file

```{r}
zip <- zip %>% 
  mutate(area_nr = parse_number(Nr)) %>% 
  mutate(zip_4 = str_extract(PLZ, "\\d{4}"))
```







```{r join-elec-with-zip}
d_zip <- d_short_nona %>% 
  left_join(zip, by = c("area_nr", "area_nr"))
```

```{r}
nrow(d_zip)
nrow(d_zip2)
```






Note that a left join was performed.


### Check


```{r}
d_zip %>% 
  filter(is.na(PLZ))
# no na
```




## Join election data with values data


VALUES PLZ with 4 digits only:


```{r}
values <- values %>% 
  mutate(zip_4 = str_extract(PLZ, "\\d{4}"))
```




```{r join-elec-with-values}

d_joined <- d_zip %>% 
  left_join(values, by= c("PLZ", "PLZ"))

d_joined2 <- d_zip %>% 
  left_join(values, by = c("zip_4", "zip_4"))
```

Note the type of join

### Check

```{r}
d_joined %>% 
  drop_na(fit.cluster, PLZ) %>% 
  nrow()

# 14k
```


How many different areas in total?

```{r}
d_joined %>% 
  summarise(n_distinct(area_nr))
```


thats'ok, 299 different areas. Go ahead.


How many areas for which we have values?


```{r}
d_joined2 <- d_joined %>% 
  drop_na(fit.cluster) %>% 
  group_by(area_nr) %>% 
  glimpse()

d_joined2 <- d_joined2 %>% ungroup()
```


There are values data for 271 districts!


## Check for Nürnberg, Stuttgart

```{r}
d_joined %>% 
  drop_na(fit.cluster) %>% 
  filter(str_detect(area_name, "Stutt"))
```


ONly 6 rows for Stuttgart, strange.

```{r}
d_joined %>% 
  drop_na(fit.cluster) %>% 
  filter(str_detect(area_name, "Nürnb"))
```



```{r}
d_joined %>% 
  drop_na(fit.cluster) %>% 
  filter(str_detect(area_name, "Stutt|Nürnb"))
```



No data for Nürnberg! Can that be?



# Join election data with zip data and values data SHORT format


## Summarise values data to Wahlkreis levels

Glimpse:

```{r}
values %>% 
  head() %>% 
  kable()
```


Compute proportion for each cluster for each wahlkreis.


Helper function:

```{r}
count_wertetypen <- function(df){
  df %>% 
    mutate(wertetyp = as.character(fit.cluster)) %>% 
    select(wertetyp, area_nr) %>% 
    # drop_na() %>% 
    count(wertetyp) %>% 
    spread(key = wertetyp, value = n) 
}
```




This function computes the proportion for each consumer type per area (plus one proportion for NA for each area).

```{r}
werte <- d_joined %>%
  group_split(area_nr) %>%  # group by Wahlkreis 
  map_dfr(count_wertetypen)  %>% 
  add_id(var = "area_nr")
```


Let's check the number of different areas.

```{r}
werte %>% 
  summarise(n_distinct(area_nr))
```


```{r}
werte %>% 
  distinct(area_nr) %>% pull()
```


Let's check for how many areas we have complete data
```{r}
werte %>% 
  drop_na()
```


Let's visualize that in a heatmap:


```{r}
p_heatmap_areas <- werte %>% 
  select(-"<NA>") %>% 
  gather(key = typ, value = n_pro_wahlkreis, -area_nr) %>% 
  ggplot(aes(x = typ, y = area_nr)) +
  geom_tile(aes(fill = n_pro_wahlkreis)) +
  scale_y_continuous(limits = c(0, 299))

p_heatmap_areas

ggsave(filename = paste0(proj_root, "/img/p_heatmap_areas.png"),
       width = 8, height = 4)
```



Now we join that to the n=299 data frame:

```{r}
d_joined_values <- d_short_nona %>% 
  full_join(werte, by = c("area_nr", "area_nr")) %>% 
  rename(type_unknown = "<NA>") %>% 
  mutate(n_wertetypen = Genieser + Harmoniesuchende + Hedonisten + 
           Selbstbestimmte + Werschätzenden + Konformisten + 
           Verantwortungsverweiterer)
```


```{r}
d_joined_values %>% 
  arrange(-n_wertetypen) %>% 
  select(n_wertetypen, area_nr, area_name)
```



# Check


```{r}
value_types <- c("Genieser", "Harmoniesuchende", "Hedonisten", 
                 "Selbstbestimmte", "Werschätzenden", 
                 "type_unknown", "Konformisten",
                 "Verantwortungsverweiterer")
```






## For how many areas do we have consumer values?


```{r}
d_joined_values %>% 
  summarise(sum(n_wertetypen, na.rm = TRUE))
```

In sum, approx. 12k.


But we have a lot of rows with *some* missing values. If we drop all such rows, only 79 remain:
```{r}
d_joined_values %>% 
  drop_na(Genieser:Verantwortungsverweiterer) %>% nrow()
```



Only 79, out of 299. That's a lot of data lost :-(.

To put it differently:

```{r}
d_joined_values %>% 
  mutate(sum_isna = rowSums(is.na(select(d_joined_values, 
                                         Genieser:Verantwortungsverweiterer)))) %>% 
  filter(sum_isna == 0) %>% 
  nrow()
```



How many rows are *completely* empty?

```{r}
d_joined_values %>% 
  dplyr::filter(n_wertetypen == 0 | is.na(n_wertetypen)) %>% 
  nrow()
```


That's at least 180! Not so bad.


Let's check for the big cities:

```{r}
dummy <- d_joined_values %>% 
  select(area_name, n_wertetypen)
```

No data for Stuttgart and Nürnberg, strange.



# Check for missing PLZs



```{r}
filter(zip, str_detect(PLZ, "2203")) %>% nrow()
filter(values, str_detect(PLZ, "2203")) %>% nrow()


filter(zip, str_detect(PLZ, "8803")) %>% nrow()
filter(values, str_detect(PLZ, "8803")) %>% nrow()

```







## Check row count


Number of rows in the 'values' dataset should approximate the row count in the values data set.

```{r}
nrow(values)/nrow(d_joined)
```


## Check effective n

```{r}
d_joined_values %>% 
  summarise(sum(n_wertetypen, na.rm = TRUE))
```




## Number of missings per column


```{r}
d_joined %>% 
  map(~ sum(is.na(.)))
```


## Number of missings per row

```{r}
d_joined %>% 
  mutate(NA_row = rowSums(is.na(.))) %>% 
  ggplot(aes(x = NA_row)) +
  geom_histogram()
```





# Write final/output data to file

The long file:

```{r}
d_joined_long_final <- d_joined

write_csv(d_joined_long_final, 
          path = "objects/d_joined_long_final.csv")
```


The short file:

```{r}
write_csv(d_joined_values,
          path = "objects/d_joined_values.csv")
```


State id dictionnary:

```{r}
save(state_dict,
     file = paste0(proj_root, "/objects/state_dict.Rda"))
```
