==> Code/02-data-preparation.Rmd <==
---
title: "Data preparation"
author: "Ich"
date: "3/9/2019"
output: html_document
editor_options: 
  chunk_output_type: console
---



```{r knitr-setup}
knitr::opts_chunk$set(
  comment = "#>",
  collapse = TRUE,
  message = FALSE,
  warning = FALSE,
  cache = TRUE,
  out.width = "70%",
  fig.align = 'center',
  fig.width = 6,
  fig.asp =  0.4,  #0.618,  # 1 / phi
  fig.show = "hold",
  size = "tiny"
)

```



# Load packages


```{r}
library(sjmisc)
library(viridis)
library(pradadata)  # elec data
library(tidyverse)
library(knitr)
library(mosaic)
library(moments)  # skewness, kurtosis

library(rethinking)  # Bayes modeling

```


# Load data


```{r}
d <- read_csv("../objects/d_joined_values.csv")
```


# Rename variables

Value type variable names:

```{r}
value_types <- c("Genieser", "Harmoniesuchende", "Hedonisten", 
                 "Selbstbestimmte", "Werschätzenden", 
                 "type_unknown", "Konformisten",
                 "Verantwortungsverweiterer")
```

Rename all seven types:

```{r}
d <- d %>% 
  rename(enjoyer = Genieser,
         harmony_seeker = Harmoniesuchende,
         hedonist = Hedonisten,
         self_determined = Selbstbestimmte,
         appreciater = Werschätzenden,
         conformist = Konformisten,
         responsibility_denier = Verantwortungsverweiterer)

attr(d, "renamed") <- TRUE

```


Here's the list as a string vector:

```{r}
consumer_types <- c("enjoyer", "harmony_seeker",
                    "hedonist", "self_determined",
                    "appreciater", "type_unknown",
                    "conformist", "responsibility_denier")
```






# Check distribution of afd_prop (outcome)

## afd_prop raw

```{r results = "hold"}
p_afd_prop_dens <- gf_dens(~afd_prop, data = d) %>% 
  gf_labs(title = paste0(" Skewness: ", round(skewness(d$afd_prop), 2),
                         "; Kurtosis: ", round(kurtosis(d$afd_prop), 2)))
gf_qqline(~afd_prop, data = d) %>% 
  gf_qq(~afd_prop)
```

```{r}
skewness(d$afd_prop)
kurtosis(d$afd_prop)
```


That's not normally distributed. Let's try to normalize it:


## afd_prop log

```{r p-log, results = "hold"}
p_afd_prop_log_dens <- gf_dens(~log(afd_prop), data = d) %>% 
  gf_labs(title = paste0(" Skewness: ", round(skewness(log(d$afd_prop)), 2),
                         "; Kurtosis: ", round(kurtosis(log(d$afd_prop)), 2)))
                         
gf_qqline(~log(afd_prop), data = d) %>% 
  gf_qq(~log(afd_prop), data = d)
```

```{r}
skewness(log(d$afd_prop))
kurtosis(log(d$afd_prop))
```

Skewness is good; kurtosis is good, since the kurtosis of the normal distribution is 3!


```{r}
library(gridExtra)

grid.arrange(p_afd_prop_dens, p_afd_prop_log_dens, nrow = 1)


```


Let's add this variable to thes datafrae:

```{r mutate-log}
d <- d %>% 
  mutate(afd_prop_log = log(afd_prop))
```


Does z-standardizing exert an impact here?


```{r}
d %>% 
  select(afd_prop_log) %>% 
  sjmisc::std() %>% 
  summarise_all(list(skewness = ~skewness(.),
                     kurtosis = ~kurtosis(.)))
```


That makes no difference.
## afd_prop sqrt

```{r results = "hold"}
gf_dens(~sqrt(afd_prop), data = d)
gf_qqline(~sqrt(afd_prop), data = d) %>% 
  gf_qq(~sqrt(afd_prop), data = d)
```


## afd_prop $x^-1$

```{r}
gf_dens(~(afd_prop^-1), data = d)
gf_qqline(~(afd_prop^-1), data = d) %>% 
  gf_qq(~(afd_prop^-1))

```


## afd_prop $x^-2$

```{r}
gf_dens(~(afd_prop^-2), data = d)
gf_qqline(~(afd_prop^-2), data = d) %>% 
  gf_qq(~(afd_prop^-2))

```




# z-standardize


```{r d-short-z}
d %>% 
  sjmisc::std() %>%  
  select(-c(state_z, area_nr_z, area_name_z, 
            state_id_z, east_z, east_num_z, 
            paste0(consumer_types, "_z"))) -> d

attr(d, "z-standardized") <- TRUE

names(d)
```


# Check distrib of z-values


```{r}
d %>% 
  gf_dens(~afd_prop_log_z)

gf_qqline(~ afd_prop_log_z, data = d) %>% 
  gf_qq(~ afd_prop_log_z)
```




```{r}
skewness(d$afd_prop_log_z)
kurtosis(d$afd_prop_log_z)
```

That's fine!


# Save data


```{r}
write_csv(d,
          path = "../objects/d_prepared.csv")
```


# sessionInfo

```{r}
sessionInfo()
```
