==> Code/03-data-modeling.Rmd <==
---
title: "Data modeling"
author: "Ich"
date: "3/9/2019"
output: html_document
params:
  readonly: FALSE
editor_options: 
  chunk_output_type: console
---



```{r knitr-setup}
knitr::opts_chunk$set(
  comment = "#>",
  collapse = TRUE,
  message = FALSE,
  warning = FALSE,
  cache = TRUE,
  out.width = "70%",
  fig.align = 'center',
  fig.width = 6,
  fig.asp =  0.4,  #0.618,  # 1 / phi
  fig.show = "hold",
  size = "tiny"
)

```



# Load packages


```{r}
library(sjmisc)
library(viridis)
library(pradadata)  # elec data
library(tidyverse)
library(knitr)
library(tidylog)
library(rprojroot)



library(rethinking)  # Bayes modeling
library("rstanarm")
```



# Setup


```{r}
proj_root <- rprojroot::find_root(as.root_criterion(is_rstudio_project))

options(mc.cores = parallel::detectCores())
```



# Load data

Raw data:

```{r}
d <- read_csv(paste0(proj_root, "/objects/d_prepared.csv"))
d <- as.data.frame(d)
```


And now the data frame with the reduced sample size of n=79,
because that's the number of complete rows for the modeling variables. 

However, we let all variables remain in the dataframe for computing prediction error later on.


```{r}
standard_model_vars <- c("afd_prop_log_z", "for_prop_z", "unemp_prop_z", "east")
consumer_types <- c("enjoyer", "harmony_seeker",
                    "hedonist", "self_determined",
                    "appreciater", "type_unknown",
                    "conformist", "responsibility_denier")
```


```{r}
d79 <- d %>% 
  drop_na(one_of(c(consumer_types, standard_model_vars))) %>% 
  as.data.frame()

dim(d)
dim(d79)
```


objects from models:

```{r eval = params$readonly}
load(paste0(proj_root, "/objects/models.Rda"))
```




In addition, we will make use of geo data in order to geoplot the modeling results. The Bundeswahlleiter provides such data (again via `pradadata`):

```{r data-shp}
data("wahlkreise_shp")
```









# Helper stuff




```{r my-traceplot}
my_traceplot <- function(model) {
  rstan::traceplot(model@stanfit, pars=names(model@start[[1]]))
}
```


# Modeling

## model output list

I use this list to store the model outputs.

```{r eval = !params$readonly}
if (!(exists("models"))) models <- list()
```


## Null models


### Normal null model (model 00)


```{r m0, eval = TRUE, results = "hide"}
d_model <- d[, c("afd_prop_log_z"), drop = FALSE]
nrow(d_model)

m0 <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),    
    mu <- alpha,
    alpha ~ dnorm(0, 1),
    sigma ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2
)
```



```{r}
precis(m0, depth = 2)
coeftab(m0)
my_traceplot(m0)
models[["m0"]] <- m0
```


### m0a: Normal null model (model 00) with reduced sample size

```{r m0a, eval = TRUE, results = "hide"}
d_model <- d79
nrow(d_model)

m0a <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),    
    mu <- alpha,
    alpha ~ dnorm(0, 1),
    sigma ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2
)
```


```{r}
precis(m0a, depth = 2)
coeftab(m0a)
my_traceplot(m0a)
models[["m0a"]] <- m0a
WAIC(m0a)
```



### m0b: Normal null model (model 00) with reduced sample size, sigma as Cauchy

```{r m0b, eval = TRUE, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars))) %>% 
  drop_na() %>% 
  as.data.frame()
nrow(d_model)

m0b <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),    
    mu <- alpha,
    alpha ~ dnorm(0, 1),
    sigma ~ dcauchy(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores = 2
)
```


```{r}
precis(m0b, depth = 2)
coeftab(m0b)
my_traceplot(m0b)
models[["m0b"]] <- m0b
WAIC(m0b)
```



## m0c: null model for afd_prop (not log, not z, just plain proportions)




```{r m0c, eval = TRUE, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "afd_prop"))) %>% 
  drop_na() %>% 
  as.data.frame()
dim(d_model)

m0c <- map2stan(
  alist(
    afd_prop ~ dnorm(mu, sigma),
    mu <- alpha,
    alpha ~ dnorm(0, 1),
    sigma ~ dcauchy(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores = 2
)
```


```{r}
precis(m0c, depth = 2)
coeftab(m0c)
my_traceplot(m0c)
models[["m0c"]] <- m0c
WAIC(m0c)
```





### m0d: As m0b, but outcome not standardized

```{r m0d, eval = TRUE, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "afd_prop_log" ))) %>% 
  drop_na() %>% 
  as.data.frame()
dim(d_model)

m0d <- map2stan(
  alist(
    afd_prop_log ~ dnorm(mu, sigma),    
    mu <- alpha,
    alpha ~ dnorm(0, 1),
    sigma ~ dcauchy(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores = 2
)
```


```{r}
precis(m0d, depth = 2)
coeftab(m0d)
my_traceplot(m0d)
models[["m0d"]] <- m0d
WAIC(m0d)
```







## m0e: null model for afd_prop POISSON




```{r m0e, eval = TRUE, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "state_id", "afd_votes"))) %>% 
  drop_na() %>% 
  as.data.frame()

dim(d_model)

m0e <- map2stan(
  alist(
    afd_votes ~ dpois(lambda), 
    
    log(lambda) <- a,
    
    a <- dnorm(0, 1)
 ),
  data = d_model,
  chains = 2,
  cores = 2,
  iter = 2000, warmup = 1500,
  control = list(adapt_delta = 0.95)
)
```


```{r}
precis(m0e, depth = 2)
coeftab(m0e)
my_traceplot(m0e)
models[["m0e"]] <- m0e
WAIC(m0c)
```








## One level models


### Model 01a: - 8 consumer vars + unemp-z + foreign-z, REDUCED sample size



```{r m1a, error = TRUE, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars))) %>% 
  drop_na() %>% 
  as.data.frame()

class(d_model)
nrow(d_model)

m1a <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),
    mu <-  beta0[east] +  beta1*for_prop_z + beta2*unemp_prop_z + 
      beta3*enjoyer + beta4*harmony_seeker + beta5*self_determined +
      beta6*appreciater + beta7*conformist + beta8*type_unknown +
      beta9*responsibility_denier,
    beta0[east] ~ dnorm(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1),
    beta3 ~ dnorm(0, 1),   
    beta4 ~ dnorm(0, 1),    
    beta5 ~ dnorm(0, 1),    
    beta6 ~ dnorm(0, 1),
    beta7 ~ dnorm(0, 1),
    beta8 ~ dnorm(0, 1),
    beta9 ~ dnorm(0, 1),
    sigma ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores = 1
)
```

Check the results:

```{r}
my_traceplot(m1a)
precis(m1a, depth = 2)
models[["m1a"]] <- m1a
WAIC(m1a)
```



### M01b: sigma cauchy distributed

Now with sigma Cauchy distributed:

```{r m1b, error = TRUE, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars))) %>% 
  drop_na() %>% 
  as.data.frame()

class(d_model)
nrow(d_model)

m1b <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),
    mu <-  beta0[east] +  beta1*for_prop_z + beta2*unemp_prop_z + 
      beta3*enjoyer + beta4*harmony_seeker + beta5*self_determined +
      beta6*appreciater + beta7*conformist + beta8*type_unknown +
      beta9*responsibility_denier,
    beta0[east] ~ dnorm(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1),
    beta3 ~ dnorm(0, 1),   
    beta4 ~ dnorm(0, 1),    
    beta5 ~ dnorm(0, 1),    
    beta6 ~ dnorm(0, 1),
    beta7 ~ dnorm(0, 1),
    beta8 ~ dnorm(0, 1),
    beta9 ~ dnorm(0, 1),
    sigma ~ dcauchy(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores = 2
)
```

Check the results:

```{r}
my_traceplot(m1b)
precis(m1b, depth = 2)
WAIC(m1b)
models[["m1b"]] <- m1b
```


### Model 02 -  unemp-z + foreign-z NO consumer vars



```{r m2, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars))) %>% 
  drop_na() %>% 
  as.data.frame()

nrow(d_model)


m2 <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),
    mu <- alpha +  beta1*for_prop_z + beta2*unemp_prop_z,
    alpha ~ dnorm(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1),
    sigma ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores = 2)
```



Check the results:

```{r}
my_traceplot(m2)
precis(m2, depth = 2)
models[["m2"]] <- m2
WAIC(m2)
```


### M02a - reduced sample size (Gaussian sigma)

Now with the reduced sample size:

```{r m2a, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars))) %>% 
  drop_na() %>% 
  as.data.frame()

nrow(d_model)


m2a <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),
    mu <- alpha +  beta1*for_prop_z + beta2*unemp_prop_z,
    alpha ~ dnorm(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1),
    sigma ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores = 2)
```



Check the results:

```{r}
my_traceplot(m2a)
precis(m2a, depth = 2)
models[["m2a"]] <- m2a
WAIC(m2a)
```

### M02b - sigma Cauchy distributed



```{r m2b, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars))) %>% 
  drop_na() %>% 
  as.data.frame()

nrow(d_model)


m2b <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),
    mu <- alpha +  beta1*for_prop_z + beta2*unemp_prop_z,
    alpha ~ dnorm(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1),
    sigma ~ dcauchy(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores = 2)
```



Check the results:

```{r}
my_traceplot(m2b)
precis(m2b, depth = 2)
models[["m2b"]] <- m2b
WAIC(m2b)
```



## Multilevel (two levels)


### M03: east + for_prop-z + unemp_prop-z

```{r m3, eval = TRUE, results = "hide"}
d_model <- d[, c("afd_prop_log_z", "for_prop_z", "unemp_prop_z", "east")]
nrow(d_model)

m3 <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),    
    mu <- beta0[east] +  beta1*for_prop_z + beta2*unemp_prop_z,
    beta0[east] ~ dnorm(0, sigma2),
    sigma ~ dnorm(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1),
    sigma2 ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores = 2
)
```



```{r}
precis(m3, depth = 2)
coeftab(m3)
my_traceplot(m3)
models[["m3"]] <- m3
```

### M03a: east + for_prop-z + unemp_prop-z REDUCED sample size

```{r m3a, eval = TRUE, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars))) %>% 
  drop_na() %>% 
  as.data.frame()

nrow(d_model)

m3a <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),    
    mu <- beta0[east] +  beta1*for_prop_z + beta2*unemp_prop_z,
    beta0[east] ~ dnorm(0, sigma2),
    sigma ~ dnorm(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1),
    sigma2 ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores = 2
)
```



```{r}
precis(m3a, depth = 2)
coeftab(m3a)
my_traceplot(m3a)
models[["m3a"]] <- m3a
WAIC(m3a)
```


### M03b: east + for_prop-z + unemp_prop-z REDUCED sample size, CAUCHY

```{r m3b, eval = TRUE, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars))) %>% 
  drop_na() %>% 
  as.data.frame()

nrow(d_model)

m3b <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),    
    mu <- beta0[east] +  beta1*for_prop_z + beta2*unemp_prop_z,
    beta0[east] ~ dnorm(0, sigma2),
    sigma ~ dcauchy(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1),
    sigma2 ~ dcauchy(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores = 2
)
```



```{r}
precis(m3b, depth = 2)
coeftab(m3b)
my_traceplot(m3b)
models[["m3b"]] <- m3b
WAIC(m3b)
```





### M03d: east + for_prop-z + unemp_prop-z, outcome afd_prop_log

Outcome not standardiezd

```{r m3b, eval = TRUE, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "afd_prop_log"))) %>% 
  drop_na() %>% 
  as.data.frame()

dim(d_model)

m3d <- map2stan(
  alist(
    afd_prop_log ~ dnorm(mu, sigma),    
    mu <- beta0[east] +  beta1*for_prop_z + beta2*unemp_prop_z,
    beta0[east] ~ dnorm(0, sigma2),
    sigma ~ dcauchy(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1),
    sigma2 ~ dcauchy(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores = 2,
  iter = 5000, warmup = 1000
)
```



```{r}
precis(m3d, depth = 2)
coeftab(m3d)
my_traceplot(m3d)
models[["m3d"]] <- m3d
WAIC(m3d)
```




### M04: state + for_prop-z + unemp_prop-z

```{r m4, eval = TRUE, results = "hide"}
d_model <- d[, c("afd_prop_log_z", "for_prop_z", "unemp_prop_z", "state_id")]
nrow(d_model)

m4 <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),    
    mu <- beta0[state_id] +  beta1*for_prop_z + beta2*unemp_prop_z,
    beta0[state_id] ~ dnorm(0, sigma2),
    sigma ~ dnorm(0, 1),
    sigma2 ~ dnorm(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores  = 2
)
```



```{r out.width="100%"}
precis(m4, depth = 2)
coeftab(m4)
my_traceplot(m4)
models[["m4"]] <- m4
```





### M04a: state + for_prop-z + unemp_prop-z, REDUCED sample size



```{r m4a, eval = TRUE, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "state_id"))) %>% 
  drop_na() %>% 
  as.data.frame()

nrow(d_model)

m4a <- map2stan(
  alist(
    afd_prop ~ dnorm(mu, sigma),    
    mu <- beta0[state_id] +  beta1*for_prop_z + beta2*unemp_prop_z,
    beta0[state_id] ~ dnorm(0, sigma2),
    sigma ~ dnorm(0, 1),
    sigma2 ~ dnorm(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores  = 2
)
```



```{r out.width="100%"}
precis(m4a, depth = 2)
coeftab(m4a)
my_traceplot(m4a)
models[["m4a"]] <- m4a

```


### M04b: state + for_prop-z + unemp_prop-z, REDUCED sample size, Cauchy



```{r m4b, eval = TRUE, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "state_id"))) %>% 
  drop_na() %>% 
  as.data.frame()

nrow(d_model)

m4b <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),    
    mu <- beta0[state_id] +  beta1*for_prop_z + beta2*unemp_prop_z,
    beta0[state_id] ~ dnorm(0, sigma2),
    sigma ~ dcauchy(0, 1),
    sigma2 ~ dcauchy(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores  = 2
)
```



```{r out.width="100%"}
precis(m4b, depth = 2)
coeftab(m4b)
my_traceplot(m4b)
models[["m4b"]] <- m4b
```



### m4b_stan: STANARM variant



```{r}

m4b_stan <- stan_glmer(afd_prop ~ for_prop_z + unemp_prop_z +
                                (1 |state_id), data = d79)


summary(m4b_stan)

rsq <- bayes_R2(m4b_stan)
print(median(rsq))  

prior_summary(object = m4b_stan)

print(m4b_stan, digits = 2)
```






### M5: state + foreing + unemp + consumer, outcome: afd_prop






```{r m5, eval = TRUE, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "state_id", "afd_prop"))) %>% 
  drop_na() %>% 
  as.data.frame()

dim(d_model)

m5 <- map2stan(
  alist(
    afd_prop ~ dnorm(mu, sigma),    
    mu <- beta0[state_id] +  beta1*for_prop_z + beta2*unemp_prop_z,
    beta0[state_id] ~ dnorm(0, sigma2),
    sigma ~ dcauchy(0, 1),
    sigma2 ~ dcauchy(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores  = 2
)
```



```{r out.width="100%"}
precis(m5, depth = 2)
coeftab(m5)
my_traceplot(m5)
models[["m5"]] <- m5

pairs(m5)
WAIC(m5)
```



### M5b: all predictors, state as multilevel, afd_prop_log_z


```{r mb5}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "state_id"))) %>% 
  drop_na() %>% 
  as.data.frame()

dim(d_model)

m5b <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),
    mu <- beta0[state_id] +  beta1*for_prop_z + beta2*unemp_prop_z +
      beta3*enjoyer + beta4*harmony_seeker + beta5*self_determined +
      beta6*appreciater + beta7*conformist + beta8*type_unknown +
      beta9*responsibility_denier,
    beta0[state_id] ~ dnorm(0, sigma2),
    sigma ~ dcauchy(0, 1),
    sigma2 ~ dcauchy(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1),
    beta3 ~ dnorm(0, 1),   
    beta4 ~ dnorm(0, 1),    
    beta5 ~ dnorm(0, 1),    
    beta6 ~ dnorm(0, 1),
    beta7 ~ dnorm(0, 1),
    beta8 ~ dnorm(0, 1),
    beta9 ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores  = 2
)
```



```{r out.width="100%"}
precis(m5b, depth = 2)
coeftab(m5b)
my_traceplot(m5b)
models[["m5b"]] <- m5b

WAIC(m5b)
```




### M6b: MAIN HYP: east (ML) + foreign + unemp + 8 consumer




```{r m6b}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "east"))) %>% 
  drop_na() %>% 
  as.data.frame()

dim(d_model)

m6b <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),
    mu <- beta0[east] +  beta1*for_prop_z + beta2*unemp_prop_z +
      beta3*enjoyer + beta4*harmony_seeker + beta5*self_determined +
      beta6*appreciater + beta7*conformist + beta8*type_unknown +
      beta9*responsibility_denier,
    beta0[east] ~ dnorm(0, sigma2),
    sigma ~ dcauchy(0, 1),
    sigma2 ~ dcauchy(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1),
    beta3 ~ dnorm(0, 1),   
    beta4 ~ dnorm(0, 1),    
    beta5 ~ dnorm(0, 1),    
    beta6 ~ dnorm(0, 1),
    beta7 ~ dnorm(0, 1),
    beta8 ~ dnorm(0, 1),
    beta9 ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores  = 2
)
```



```{r out.width="100%"}
precis(m6b, depth = 2)
coeftab(m6b)
my_traceplot(m6b)
models[["m6b"]] <- m6b

WAIC(m6b)
```









### M6d: MAIN HYP: east (ML) + foreign + unemp + 8 consumer, afd_prop_log as outcome 

OUtcome is not z-standardized


```{r m6b}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "east", "afd_prop_log"))) %>% 
  drop_na() %>% 
  as.data.frame()

dim(d_model)

m6d <- map2stan(
  alist(
    afd_prop_log ~ dnorm(mu, sigma),
    mu <- beta0[east] +  beta1*for_prop_z + beta2*unemp_prop_z +
      beta3*enjoyer + beta4*harmony_seeker + beta5*self_determined +
      beta6*appreciater + beta7*conformist + beta8*type_unknown +
      beta9*responsibility_denier,
    beta0[east] ~ dnorm(0, sigma2),
    sigma ~ dcauchy(0, 1),
    sigma2 ~ dcauchy(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1),
    beta3 ~ dnorm(0, 1),   
    beta4 ~ dnorm(0, 1),    
    beta5 ~ dnorm(0, 1),    
    beta6 ~ dnorm(0, 1),
    beta7 ~ dnorm(0, 1),
    beta8 ~ dnorm(0, 1),
    beta9 ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores  = 2
)
```



```{r out.width="100%"}
precis(m6d, depth = 2)
coeftab(m6d)
my_traceplot(m6d)
models[["m6d"]] <- m6d

WAIC(m6d)
```






### M7b:  8 consumer only




```{r m7b}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "east"))) %>% 
  drop_na() %>% 
  as.data.frame()

dim(d_model)

m7b <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),
    mu <- beta3*enjoyer + beta4*harmony_seeker + beta5*self_determined +
      beta6*appreciater + beta7*conformist + beta8*type_unknown +
      beta9*responsibility_denier,

    sigma ~ dcauchy(0, 1),

    beta3 ~ dnorm(0, 1),   
    beta4 ~ dnorm(0, 1),    
    beta5 ~ dnorm(0, 1),    
    beta6 ~ dnorm(0, 1),
    beta7 ~ dnorm(0, 1),
    beta8 ~ dnorm(0, 1),
    beta9 ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores  = 2
)
```



```{r out.width="100%"}
precis(m7b, depth = 2)
coeftab(m7b)
my_traceplot(m7b)
models[["m7b"]] <- m7b

WAIC(m7b)
```




### M8b:  SEVEN consumer variables only, no NA-variable




```{r m8b}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "east"))) %>% 
  drop_na() %>% 
  select(-type_unknown) %>% 
  as.data.frame()

dim(d_model)

m8b <- map2stan(
  alist(
    afd_prop_log_z ~ dnorm(mu, sigma),
    mu <- beta3*enjoyer + beta4*harmony_seeker + beta5*self_determined +
      beta6*appreciater + beta7*conformist +
      beta9*responsibility_denier,

    sigma ~ dcauchy(0, 1),

    beta3 ~ dnorm(0, 1),   
    beta4 ~ dnorm(0, 1),    
    beta5 ~ dnorm(0, 1),    
    beta6 ~ dnorm(0, 1),
    beta7 ~ dnorm(0, 1),
    
    beta9 ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores  = 2
)
```



```{r out.width="100%"}
precis(m8b, depth = 2)
coeftab(m8b)
my_traceplot(m8b)
models[["m8b"]] <- m8b

WAIC(m8b)
```




### M9b: As M4b, but count model (Poisson)






```{r m9b}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "state_id", "afd_votes"))) %>% 
  drop_na() %>% 
  as.data.frame()

dim(d_model)

m9b <- map2stan(
  alist(
    afd_votes ~ dpois(lambda), 
    
    log(lambda) <- beta0[state_id] +  beta1*for_prop_z + beta2*unemp_prop_z,
    
    beta0[state_id] ~ dnorm(0, sigma2),
    
    sigma2 ~ dcauchy(0, 1),
    
    c(beta1, beta2) ~ dnorm(0, 1)
 ),
  data = d_model,
  chains = 2,
  cores  = 2
)


```




```{r out.width="100%"}

precis(m9b, depth = 2)
coeftab(m9b)
my_traceplot(m9b)
models[["m9b"]] <- m9b

WAIC(m9b)
```




### M10c: As M4b, but not z-standardized outcome




```{r m10c, eval = TRUE, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "state_id", "afd_prop_log"))) %>% 
  drop_na() %>% 
  as.data.frame()

dim(d_model)


str(d_model$state)

m10c <- map2stan(
  alist(
    afd_prop_log ~ dnorm(mu, sigma),    
    mu <- beta0[state_id] +  beta1*for_prop_z + beta2*unemp_prop_z,
    beta0[state_id] ~ dnorm(0, sigma2),
    sigma ~ dcauchy(0, 1),
    sigma2 ~ dcauchy(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1)
  ),
  data = d_model,
  chains = 2,
  cores = 2,
  iter = 5000, warmup = 3000,
  control = list(adapt_delta = 0.99,
                 max_treedepth = 15)
)
```



```{r out.width="100%"}
precis(m10c, depth = 2)
coeftab(m10c)
my_traceplot(m10b)
models[["m10c"]] <- m10c
```






### M11d: As M5b, but not z-standardized outcome




```{r m11d, eval = TRUE, results = "hide"}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "state_id", "afd_prop_log"))) %>% 
  drop_na() %>% 
  as.data.frame()

dim(d_model)

m11d <- map2stan(
  alist(
    afd_prop_log ~ dnorm(mu, sigma),  
    
    mu <- beta0[state_id] +  beta1*for_prop_z + beta2*unemp_prop_z +
      beta3*enjoyer + beta4*harmony_seeker + beta5*self_determined +
      beta6*appreciater + beta7*conformist + beta8*type_unknown +
      beta9*responsibility_denier,
    beta0[state_id] ~ dnorm(0, sigma2),
    sigma ~ dcauchy(0, 1),
    sigma2 ~ dcauchy(0, 1),
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1),
    beta3 ~ dnorm(0, 1),   
    beta4 ~ dnorm(0, 1),    
    beta5 ~ dnorm(0, 1),    
    beta6 ~ dnorm(0, 1),
    beta7 ~ dnorm(0, 1),
    beta8 ~ dnorm(0, 1),
    beta9 ~ dnorm(0, 1)
  ),
  
  data = d_model,
  chains = 2,
  cores = 2,
  iter = 2000, warmup = 1000,
  control = list(adapt_delta = 0.95,
                 max_treedepth = 15)
)
```



```{r out.width="100%"}
precis(m11d, depth = 2)
coeftab(m11d)
my_traceplot(m11d)
models[["m11d"]] <- m11d


m11d
```






### M12d: As M9b, but with consumer types






```{r m12d}
d_model <- d %>% 
  select(one_of(c(consumer_types, standard_model_vars, "state_id", "afd_votes"))) %>% 
  drop_na() %>% 
  as.data.frame()

dim(d_model)

m12d <- map2stan(
  alist(
    afd_votes ~ dpois(lambda), 
    
    log(lambda) <- beta0[state_id] +  beta1*for_prop_z + beta2*unemp_prop_z +
      beta3*enjoyer + beta4*harmony_seeker + beta5*self_determined +
      beta6*appreciater + beta7*conformist + beta8*type_unknown +
      beta9*responsibility_denier,
    
    beta0[state_id] ~ dnorm(0, sigma2),
    
    sigma ~ dcauchy(0, 1),
    sigma2 ~ dcauchy(0, 1),
    
    beta1 ~ dnorm(0, 1),
    beta2 ~ dnorm(0, 1),
    beta3 ~ dnorm(0, 1),   
    beta4 ~ dnorm(0, 1),    
    beta5 ~ dnorm(0, 1),    
    beta6 ~ dnorm(0, 1),
    beta7 ~ dnorm(0, 1),
    beta8 ~ dnorm(0, 1),
    beta9 ~ dnorm(0, 1)    
  ),
  data = d_model,
  chains = 4,
  cores = 2,
  iter = 2000, warmup = 1000,
  control = list(adapt_delta = 0.95,
                 max_treedepth = 15)
)


```




```{r out.width="100%"}

precis(m12d, depth = 2)
coeftab(m12d)
my_traceplot(m12d)
models[["m12d"]] <- m12d

WAIC(m12d)
```





### m4b_stan: STANARM variant



```{r}

m4b_stan <- stan_glmer(afd_prop ~ for_prop_z + unemp_prop_z +
                                (1 |state_id), data = d79)


summary(m4b_stan)

rsq <- bayes_R2(m4b_stan)
print(median(rsq))  

prior_summary(object = m4b_stan)

print(m4b_stan, digits = 2)
```








# Save results

```{r}
names(models)
length(names(models))
```


```{r}
save(models, 
     file = paste0(proj_root, "/objects/models.Rda"))

write.csv(d79,
          file = paste0(proj_root, "/objects/d79.csv"))
```
